+++ core/jobs/src/main/java/com/cognifide/aet/job/common/modifiers/executejavascript/ExecuteJavaScriptModifierFactory.java
@@ -21,12 +21,10 @@ import com.cognifide.aet.job.api.collector.CollectorProperties;
 import com.cognifide.aet.job.api.collector.WebCommunicationWrapper;
 import com.cognifide.aet.job.api.exceptions.ParametersException;
 import java.util.Map;
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.Reference;
-import org.apache.felix.scr.annotations.Service;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 @Component
-@Service
 public class ExecuteJavaScriptModifierFactory implements CollectorFactory {
 
   @Reference