+++ core/worker/src/main/java/com/cognifide/aet/worker/listeners/AbstractTaskMessageListener.java
@@ -18,45 +18,20 @@ package com.cognifide.aet.worker.listeners;
 import com.cognifide.aet.communication.api.queues.JmsConnection;
 import com.cognifide.aet.queues.JmsUtils;
 import com.cognifide.aet.worker.results.FeedbackQueue;
-import java.util.Map;
 import javax.jms.JMSException;
 import javax.jms.MessageConsumer;
 import javax.jms.MessageListener;
 import javax.jms.Session;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.Property;
 
-@Component
 public abstract class AbstractTaskMessageListener implements MessageListener {
 
-  protected static final String LISTENER_NAME = "name";
+  private Session jmsSession;
 
-  protected static final String CONSUMER_QUEUE_NAME = "consumerQueueName";
+  private MessageConsumer consumer;
 
-  protected static final String PRODUCER_QUEUE_NAME = "producerQueueName";
+  FeedbackQueue feedbackQueue;
 
-  protected static final String PREFETCH_SIZE_NAME = "pf";
-
-  protected static final String PREFETCH_SIZE_DEFAULT_VALUE = "1";
-
-  protected Session jmsSession;
-
-  protected MessageConsumer consumer;
-
-  protected FeedbackQueue feedbackQueue;
-
-  @Property(name = PREFETCH_SIZE_NAME, label = "Prefetch size", description = "http://activemq.apache.org/what-is-the-prefetch-limit-for.html", value = PREFETCH_SIZE_DEFAULT_VALUE)
-  private String prefetchSize;
-
-  protected void doActivate(Map<String, String> properties) {
-    setName(properties.get(LISTENER_NAME));
-    this.prefetchSize = StringUtils.defaultString(properties.get(PREFETCH_SIZE_NAME),
-        PREFETCH_SIZE_DEFAULT_VALUE);
-    String consumerQueueName = properties.get(CONSUMER_QUEUE_NAME);
-    setConsumerQueueName(consumerQueueName);
-    String producerQueueName = properties.get(PRODUCER_QUEUE_NAME);
-    setProducerQueueName(producerQueueName);
+  void doActivate(String consumerQueueName, String producerQueueName, String prefetchSize) {
 
     String queueName = consumerQueueName + "?consumer.prefetchSize=" + prefetchSize;
     try {
@@ -69,19 +44,13 @@ public abstract class AbstractTaskMessageListener implements MessageListener {
     }
   }
 
-  protected void doDeactivate() {
+  void doDeactivate() {
     if (feedbackQueue != null) {
       feedbackQueue.close();
     }
     JmsUtils.closeQuietly(consumer);
     JmsUtils.closeQuietly(jmsSession);
   }
-
-  protected abstract void setName(String name);
-
-  protected abstract void setConsumerQueueName(String consumerQueueName);
-
-  protected abstract void setProducerQueueName(String producerQueueName);
 
   protected abstract JmsConnection getJmsConnection();